# Modify a hash in place and freeze strings to reduce allocations.