@jrafanie 2015-11-17
I ran MiqReport#generate_table in ManageIQ on some of the really slow reports and there is some savings in object allocations.

## Before:
```
Time elapsed |      RSS change |       Total RSS | Allocated Objects |                         Description
 22.481769 |       146337792 |       341491712 |        15235399 |           33 - Host Summary for VMs
 33.970649 |        29888512 |       371380224 |        21535819 |       53 - Folder VMs Relationships
 17.848895 |         1949696 |       373329920 |        11623846 |  60 - Summary - VMs migration ready
 15.259537 |        19046400 |       392376320 |         9986093 | 62 - Detailed - VMs migration ready
 19.187577 |            4096 |       392380416 |        13005036 |               88 - VM Relationships
 45.467005 |         2138112 |       394518528 |        29091805 |    89 - Folder to VMs Relationships
 22.587697 |       130605056 |       525123584 |        13494862 | 104 - All Departments with Performance
  11.07995 |       -31735808 |       493387776 |         8015645 |         107 - Top Storage Consumers
```

## After:
```
 Time elapsed |      RSS change |       Total RSS | Allocated Objects |                         Description
 22.646629 |       121155584 |       315072512 |        15073316 |           33 - Host Summary for VMs
 34.076176 |        33177600 |       348250112 |        21451820 |       53 - Folder VMs Relationships
 18.487667 |         9924608 |       358174720 |        11515845 |  60 - Summary - VMs migration ready
 15.227479 |        29134848 |       387309568 |         9830093 | 62 - Detailed - VMs migration ready
 19.270113 |         -954368 |       386355200 |        12879036 |               88 - VM Relationships
 45.241741 |          835584 |       387190784 |        28923805 |    89 - Folder to VMs Relationships
 19.746829 |        81096704 |       468287488 |        13008670 | 104 - All Departments with Performance
 11.110456 |         9179136 |       477466624 |         7891707 |         107 - Top Storage Consumers
```