@jrafanie 2015-11-17
Script for prior comment:

```ruby
  def log_block_memory_allocations(description = nil)
    require "miq-process"

    $run ||= begin
      puts "Ruby: #{RUBY_VERSION}"
      puts "%10s | %15s | %15s | %15s | %35s" %
        [
          "Time elapsed",
          "RSS change",
          "Total RSS",
          "Allocated Objects",
          "Description"
        ]
      true
    end

    allocated_before = GC.stat[:total_allocated_objects]
    mem_before       = MiqProcess.processInfo[:memory_usage]
    start            = Time.now

    yield

    allocated_after = GC.stat[:total_allocated_objects]
    mem_after       = MiqProcess.processInfo[:memory_usage]
    elapsed         = Time.now - start

    puts "%10s | %15s | %15s | %15s | %35s" %
      [
          elapsed,
          (mem_after - mem_before),
          mem_after,
          (allocated_after - allocated_before),
          description
      ]
  end
end

MiqReport.order(:id).where(:id => [33, 53, 60, 62, 88, 89, 104, 107]).each do |r|
  log_block_memory_allocations("#{r.id} - #{r.name}") do
    begin
      r.generate_table
    rescue => e
      # Swallow for now
      # puts "#{r.name} #{r.id} failed with #{e}"
    end
  end
end
```