@jrafanie 2015-11-17
@Fryguy @kbrock @gtanzillo please review at your leisure, thanks!