@jrafanie 2015-11-18
> This was a great find.

Thanks... Unfortunately, inefficient cpu and memory is still much faster than database queries so there isn't a huge speed improvement here although extra churn on the ruby 2.1+ generational GC can easily cause memory to grow like crazy, especially if there are any objects that make it to the old generation and won't get GC'd until the next full GC.